unifying user controller
