+++ src/main/java/com/alkemy/ong/auth/config/SecurityConfiguration.java
@@ -1,8 +1,7 @@
 package com.alkemy.ong.auth.config;
 
 import com.alkemy.ong.auth.filter.JwtRequestFilter;
-import com.alkemy.ong.auth.service.UserDetailsCustomService;
-import io.jsonwebtoken.Jwt;
+import com.alkemy.ong.auth.service.impl.UserDetailsCustomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -14,7 +13,6 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
