+++ src/main/java/com/alkemy/ong/auth/filter/JwtRequestFilter.java
@@ -1,11 +1,10 @@
 package com.alkemy.ong.auth.filter;
 
 import com.alkemy.ong.auth.service.JwtUtils;
-import com.alkemy.ong.auth.service.UserDetailsCustomService;
+import com.alkemy.ong.auth.service.impl.UserDetailsCustomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
@@ -16,7 +15,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Collections;
 
 @Component
 public class JwtRequestFilter extends OncePerRequestFilter {
