--- src/main/java/com/alkemy/ong/auth/service/impl/UserServiceImpl.java
@@ -1,33 +1,0 @@
-package com.alkemy.ong.auth.service.impl;
-
-import com.alkemy.ong.auth.dto.UserResponseDto;
-import com.alkemy.ong.auth.repository.UserRepository;
-import com.alkemy.ong.auth.service.IUserService;
-import com.alkemy.ong.exception.NotFoundException;
-import com.alkemy.ong.mapper.UserMapper;
-import com.alkemy.ong.model.UserEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class UserServiceImpl implements IUserService {
-
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private UserMapper userMapper;
-
-    @Transactional(readOnly = true)
-    @Override
-    public List<UserResponseDto> getAllUsers() {
-        List<UserEntity> entities = this.userRepository.findAll();
-        if(entities.isEmpty()){
-            throw new NotFoundException("The list is empty");
-        }
-        List<UserResponseDto> userResponseDtos = userMapper.userEntityList2ResponseDtoList(entities,true);
-        return userResponseDtos;
-    }
-}
