--- src/main/java/com/alkemy/ong/controller/UserController.java
@@ -1,31 +1,0 @@
-package com.alkemy.ong.controller;
-
-import com.alkemy.ong.auth.dto.UserResponseDto;
-import com.alkemy.ong.auth.repository.UserRepository;
-import com.alkemy.ong.auth.service.IUserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@PreAuthorize("authenticated")
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-    @Autowired
-    private IUserService userService;
-
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @GetMapping
-    public ResponseEntity<List<UserResponseDto>> getAllUsers() {
-        List<UserResponseDto> usersDTOs = this.userService.getAllUsers();
-        return ResponseEntity.ok().body(usersDTOs);
-    }
-
-
-}
